Enhancement : change Duke class to Dude
